python/instrumentation/openinference-instrumentation-bedrock-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openinference-instrumentation-bedrock-agent"
dynamic = ["version"]
description = "OpenInference Bedrock Agent Instrumentation"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9, <3.14"
authors = [
  { name = "OpenInference Authors", email = "oss@arize.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "opentelemetry-api",
  "opentelemetry-instrumentation",
  "opentelemetry-semantic-conventions",
  "openinference-instrumentation>=0.1.17",
  "openinference-semantic-conventions>=0.1.7",
  "wrapt",
  "typing-extensions",
  "dacite>=1.8.1",
]

[project.optional-dependencies]
instruments = [
  "boto3 >= 1.28.57",
]
test = [
  "boto3 == 1.34.116",
  "opentelemetry-sdk",
  "opentelemetry-instrumentation-httpx",
  "pytest-vcr>=1.0.2",
]

[project.urls]
Homepage = "https://github.com/Arize-ai/openinference/tree/main/python/instrumentation/openinference-instrumentation-bedrock-agent"

[tool.hatch.version]
path = "src/openinference/instrumentation/bedrock_agent/version.py"

[tool.hatch.build.targets.sdist]
include = [
  "/src",
  "/tests",
]

[tool.hatch.build.targets.wheel]
packages = ["src/openinference"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = [
  "tests",
]

[tool.mypy]
strict = true
explicit_package_bases = true
exclude = [
  "examples",
  "dist",
  "sdist",
]

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
  "wrapt",
  "boto3",
  "botocore.*",
]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = ["E402", "E501"]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.isort]
force-single-line = false
